requirements.txt
gradio>=4.0.0
langchain-community>=0.0.0
sentence-transformers>=2.2.0
pandas>=2.0.0
numpy>=1.24.0
sqlite-vec>=0.1.0
openai>=1.0.0
pyprojroot>=0.2.0
pyyaml>=6.0